Move CheckoutV2 to a separate component and fixed container spacing and background.
