+++ client/a8c-for-agencies/sections/client/primary/checkout-v2/index.tsx
@@ -1,0 +1,144 @@
+import { RazorpayHookProvider } from '@automattic/calypso-razorpay';
+import { StripeHookProvider } from '@automattic/calypso-stripe';
+import { CheckoutErrorBoundary } from '@automattic/composite-checkout';
+import { createRequestCartProduct, useShoppingCart } from '@automattic/shopping-cart';
+import debugFactory from 'debug';
+import { useTranslate } from 'i18n-calypso';
+import { useEffect, useState } from 'react';
+import useProductsById from 'calypso/a8c-for-agencies/sections/marketplace/hooks/use-products-by-id';
+import { getClientReferralQueryArgs } from 'calypso/a8c-for-agencies/sections/marketplace/lib/get-client-referral-query-args';
+import { getStripeConfiguration, getRazorpayConfiguration } from 'calypso/lib/store-transactions';
+import CalypsoShoppingCartProvider from 'calypso/my-sites/checkout/calypso-shopping-cart-provider';
+import CheckoutMain from 'calypso/my-sites/checkout/src/components/checkout-main';
+import { useSelector } from 'calypso/state';
+import { getCurrentUserLocale } from 'calypso/state/current-user/selectors';
+import useFetchClientReferral from '../../hooks/use-fetch-client-referral';
+
+import './style.scss';
+
+const debug = debugFactory( 'client:checkout-v2' );
+
+/**
+ * Client Checkout Component using the WordPress.com checkout
+ */
+function ClientCheckoutContent() {
+	const translate = useTranslate();
+	const [ isReady, setIsReady ] = useState( false );
+	const [ error, setError ] = useState< string | null >( null );
+
+	const { data: referral } = useFetchClientReferral( getClientReferralQueryArgs() );
+	const { referredProducts } = useProductsById( referral?.products ?? [] );
+
+	// Access the shopping cart API
+	const { addProductsToCart, responseCart } = useShoppingCart( 'no-site' );
+
+	// Add products to cart when referral data is loaded
+	useEffect( () => {
+		// Skip if we're already ready or have an error
+		if ( isReady || error ) {
+			debug( '[A4A Checkout] Already ready or has error' );
+			return;
+		}
+
+		if ( ! referredProducts || referredProducts.length === 0 ) {
+			debug( '[A4A Checkout] No referred products' );
+			return;
+		}
+
+		debug( '[A4A Checkout] Referral', referral );
+		debug( '[A4A Checkout] Referred products', referredProducts );
+
+		const productsToAdd = referredProducts.map( ( product ) => {
+			return createRequestCartProduct( {
+				product_id: product.product_id,
+				product_slug: product.slug,
+				extra: { isA4ASitelessCheckout: true },
+			} );
+		} );
+
+		debug( '[A4A Checkout] Products to add', productsToAdd );
+
+		// Add products to cart
+		if ( productsToAdd.length > 0 ) {
+			addProductsToCart( productsToAdd )
+				.then( () => {
+					debug( '[A4A Checkout] Products added to cart successfully' );
+					debug( '[A4A Checkout] Cart', responseCart );
+					setIsReady( true );
+				} )
+				.catch( ( err ) => {
+					debug( '[A4A Checkout] Failed to add products to cart:', err );
+					setError( 'Failed to add products to cart' );
+				} );
+		} else {
+			debug( '[A4A Checkout] No matching products found to add to cart' );
+			setError( 'Could not find the requested products' );
+		}
+	}, [ isReady, error, referredProducts ] );
+
+	// Debugging: Set a timeout to force showing the checkout after 10 seconds
+	useEffect( () => {
+		if ( isReady || error ) {
+			return;
+		}
+
+		const timeoutId = setTimeout( () => {
+			debug( '[A4A Checkout] Timeout reached, showing checkout anyway' );
+			setIsReady( true );
+		}, 10000 );
+
+		return () => clearTimeout( timeoutId );
+	}, [ isReady, error ] );
+
+	// Debugging: Show loading state
+	if ( ! isReady && ! error ) {
+		return (
+			<div className="client-checkout-v2__loading">
+				<h2>{ translate( 'Loading checkout' ) }</h2>
+				<p>{ translate( 'Setting up your products for purchase' ) }</p>
+			</div>
+		);
+	}
+
+	// Debugging: Show error state
+	if ( error ) {
+		return (
+			<div className="client-checkout-v2__error">
+				<h2>{ translate( 'Error' ) }</h2>
+				<p>{ error }</p>
+			</div>
+		);
+	}
+
+	// Show checkout
+	return (
+		<div className="client-checkout-v2">
+			<CheckoutMain
+				sitelessCheckoutType="a4a"
+				redirectTo="/client/subscriptions"
+				isInModal={ false }
+				siteSlug=""
+				siteId={ 0 }
+			/>
+		</div>
+	);
+}
+
+export default function ClientCheckoutV2() {
+	const translate = useTranslate();
+	const locale = useSelector( getCurrentUserLocale );
+
+	return (
+		<CheckoutErrorBoundary
+			errorMessage={ translate( 'Sorry, there was an error loading the checkout page.' ) }
+		>
+			<CalypsoShoppingCartProvider shouldShowPersistentErrors>
+				<StripeHookProvider fetchStripeConfiguration={ getStripeConfiguration } locale={ locale }>
+					<RazorpayHookProvider fetchRazorpayConfiguration={ getRazorpayConfiguration }>
+						<ClientCheckoutContent />
+					</RazorpayHookProvider>
+				</StripeHookProvider>
+			</CalypsoShoppingCartProvider>
+		</CheckoutErrorBoundary>
+	);
+}
