+++ client/a8c-for-agencies/sections/client/primary/checkout-v2/style.scss
@@ -1,0 +1,6 @@
+.theme-a8c-for-agencies:has(.client-checkout-v2) {
+	.layout.has-no-masterbar .layout__content {
+		padding: 0;
+		background-color: var(--color-surface);
+	}
+}
